requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.1.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonformer==0.12.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.77
langchain-google-genai==2.0.10
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.32
lightning-utilities==0.15.2
loguru==0.7.3
MarkupSafe==3.0.3
marshmallow==3.26.1
faiss-cpu==1.7.4
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.6.0
networkx==3.4.2
numpy<2
openai==2.1.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.1
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.1.99
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
streamlit==1.50.0
sympy==1.14.0
tenacity==8.2.2
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers>=0.22.0
toml==0.10.2
torch==2.8.0
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
transformers>=4.56.1
trl==0.23.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wikipedia==1.4.0
Wikipedia-API==0.8.1
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.20.1
zstandard==0.25.0
pytest==8.3.3
